refactor(linter): move `oxlint` application code into separate module (#13745)
Pure refactor. Move code for running the linter in `oxlint` into a separate `run.rs` module. This is preparation for making `oxlint` into a NAPI package (#13723).
